stile: изменила тип данных id для User и Film на Long

+++ src/main/java/ru/yandex/practicum/filmorate/controllers/FilmController.java
@@ -55,7 +55,7 @@ public class FilmController {
         if (Integer.parseInt(id) <= 0) {
             throw new FilmNotFoundException(Constants.FILM_ID_INCORRECT);
         }
-        return filmStorage.getFilmById(Integer.parseInt(id));
+        return filmStorage.getFilmById(Long.parseLong(id));
     }
 
     // Добавление лайка к фильму
