+++ src/main/java/ru/yandex/practicum/filmorate/model/Id.java
@@ -13,8 +13,8 @@ public class Id {
 
     //генерирует новый id на основании последнего созданного Id
     public static Long getId(Set<Long> id) {
-        Comparator<Integer> comparator = Comparator.comparingInt(id2 -> id2);
-        LinkedList<Integer> allId = new LinkedList<>(id);
+        Comparator<Long> comparator = Comparator.comparingLong(id2 -> id2);
+        LinkedList<Long> allId = new LinkedList<>(id);
         allId.sort(comparator);
         if (allId.isEmpty()) {
             return ONE;
